Move AMP to new class/file, rename to setStorage, enum for get/set.
